indiegame.scm: comment out 'literal-script/style'
literal-script/style is not yet in the official chicken-install version of
awful, which we recommend people to use for the trainingquest to run the
site locally.  Since we don't actually use this feature, we will comment
it out for now, and uncomment it when it is available in the released
version of awful.